refactor: renamed all Python files to follow a common naming convention. Necessary refactors were made in the affected files
